shipyard_blueprints/s3/pyproject.toml
[tool.poetry]
name = "shipyard-s3"
version = "0.1.3a0"
description = "A local client for connecting and working with AWS S3"
authors = ["wrp801 <wespoulsen@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "shipyard_s3"}]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "1.34.44"
# shipyard-templates = "0.6.2a1"
shipyard-bp-utils = "1.0.2a1"

[tool.poetry.group.dev.dependencies]
# shipyard-bp-utils = {path = "../shipyard-bp-utils", develop = true}
shipyard-templates = {path = "../../shipyard-templates", develop = true}
pytest = "^7.4.0"
python-dotenv = "^1.0.1"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
